form appears in the modal. Needs to be attached handleSubmit to modals closing or modal ok to handle submit switching to redux branch
